Rework janky member profile page
The earlier implementation of the member pages was trying to replicate API like behavior by using the 404.html page to redirect to "API Endpoint". Relying on such behavior is extremely janky and especially difficult to deal with using relative addressing. Hence, I have migrated this functionality to a dedicated page `/about/member.html` which loads a member profile based on the `user` query parameter.

Will update the 404.html page to be an actual 404.html page in the future.